Add Colemak keyboard definition files
Colemak is unusual in that it doesn't have a Caps Lock key.
This means we don't need to generate capslock sections in the keymap
file. We add a flag to the keymap file [General] section to indicate
these sections are missing.

A change is also made to the way the Dvorak keymap file (00010409) is
generated, for consistency with Colemak.